// src/features/posts/types.ts
export interface IPost {
  by: string;
  descendants: number;
  id: number;
  kids: number[];
  score: number;
  time: number;
  title: string;
  type: string;
  url: string;
  userCarma: number;
}
export interface IResponse {
  type: string;
  payload: IPost;
}
export type TCustom = IResponse | [];